+++ ChangeLog.md
@@ -11,7 +11,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Added
 
-- added ...
+- added function str/quote
+- added function bytebuf-from-string, bytebuf-to-string
 
 ### Removed